requirements.txt
streamlit
python-pptx
Pillow
moviepy
yt_dlp